crysllmGen/requirements_safe.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
ase
attrs
bibtexparser
boto3
botocore
certifi
charset-normalizer
click
contourpy
cycler
dill
dnspython
einops
filelock
fonttools
frozenlist
fsspec
gitdb
GitPython
googledrivedownloader
hf-xet
hydra-core
idna
Jinja2
jmespath
joblib
kiwisolver
lightning
lightning-utilities
MarkupSafe
matgl
matminer
matplotlib
monty
mpmath
multidict
multiprocess
narwhals
networkx
numpy
omegaconf
orjson
p_tqdm
packaging
palettable
pandarallel
pandas
pathos
pillow
platformdirs
plotly
pox
ppft
propcache
protobuf
psutil
pyarrow
pydantic
pydantic_core
pymatgen
pymongo
pyparsing
python-dateutil
python-louvain
pytorch-lightning
pytz
PyYAML
rdflib
regex
requests
ruamel.yaml
ruamel.yaml.clib
s3transfer
scikit-learn
scipy
sentencepiece
sentry-sdk
setuptools
six
SMACT
smmap
spglib
sympy
tabulate
threadpoolctl
torchdata
torchmetrics
tqdm
triton
typing-inspection
typing_extensions
tzdata
uncertainties
urllib3
wandb
wheel
xxhash
yarl
